feat: shape tool
Create a new tool shape that allows the player to draw shapes on the map
by selecting a centre and vertex point. A GUI is created when the tool
is selected. At this stage only a render is drawn and no tiles are
created
Also refactor the existing entitiy painter to "painter" and split GUI